+++ torch/_dynamo/testing.py
@@ -3,12 +3,15 @@ import dis
 import functools
 import logging
 import os.path
+import random
 import re
 import sys
 import types
 import unittest
 from typing import Sequence, Union
 from unittest.mock import patch
+
+import numpy as np
 
 import torch
 from torch import fx
@@ -363,3 +366,13 @@ def expectedFailureCodegenDynamic(fn):
 def expectedFailureDynamicWrapper(fn):
     fn._expected_failure_dynamic_wrapper = True
     return fn
+
+
+def reset_rng_state(use_xla=False):
+    torch.manual_seed(1337)
+    random.seed(1337)
+    np.random.seed(1337)
+    if use_xla:
+        import torch_xla.core.xla_model as xm
+
+        xm.set_rng_state(1337, str(xm.xla_device()))
